refactor: type 및 api 분리 (#236)
* rafactor
- my-bundles type 폴더 분리 및 type명 구체화
- type 필드명 카멜케이스로 변경

* rafactor
- my-bundle api 폴더 분리
- endpoint 생성 함수명 수정

* refactor: api 요청에 필요한 파라미터 타입 분리

* refactor
- endpoint 생성 함수명 수정
- type명 수정 및 분리

* refactor: 배달부 설정 로직 수정

* style: viewport 설정

* style: eslint 적용

* refactor: 에러 처리 로직 추가

* refactor: 함수명 수정

* refactor: console.log 및 alert -> toast로 변경
