+++ src/queries/useBundlesPreviewQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { getBundlesPreview } from "@/api/bundle/api";
+import { getBundlesPreview } from "@/api/my-bundle/api";
 
 export const useBundlesPreviewQuery = () => {
   return useQuery({
