+++ src/queries/useCreateBundleMutation.ts
@@ -3,6 +3,7 @@
 import { useMutation } from "@tanstack/react-query";
 
 import { createBundle } from "@/api/bundle/api";
+import { toast } from "@/hooks/use-toast";
 import {
   useBundleNameStore,
   useSelectedBagStore,
@@ -30,6 +31,12 @@ export const useCreateBundleMutation = () => {
       if (res?.gifts?.length) {
         updateGiftBoxesFromResponse(res.gifts);
       }
+    },
+    onError: () => {
+      toast({
+        title: "보따리 생성에 실패했어요.",
+        description: "다시 시도해 주세요.",
+      });
     },
   });
 };
