+++ src/queries/useDeleteMyBundleMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
-import { deleteMyBundle } from "@/api/bundle/api";
+import { deleteMyBundle } from "@/api/my-bundle/api";
 import { toast } from "@/hooks/use-toast";
 
 export const useDeleteMyBundleMutation = () => {
@@ -14,7 +14,7 @@ export const useDeleteMyBundleMutation = () => {
     onError: () => {
       toast({
         title: "보따리를 삭제하는데 실패했어요.",
-        description: "다시 시도해주세요.",
+        description: "다시 시도해 주세요.",
       });
     },
   });
