+++ src/queries/useDeleteUserMutation.ts
@@ -1,10 +1,10 @@
 "use client";
 
 import { useMutation } from "@tanstack/react-query";
-
-import { toast } from "@/hooks/use-toast";
-import { deleteUser } from "@/api/kakao/api";
 import { useRouter } from "next/navigation";
+
+import { deleteUser } from "@/api/kakao/api";
+import { toast } from "@/hooks/use-toast";
 import { deleteToken } from "@/utils/tokenUtils";
 
 export const useDeleteUserMutation = () => {
@@ -19,7 +19,7 @@ export const useDeleteUserMutation = () => {
     onError: () => {
       toast({
         title: "회원 탈퇴에 실패했어요.",
-        description: "다시 시도해주세요.",
+        description: "다시 시도해 주세요.",
       });
     },
   });
