+++ src/queries/useGiftDetailQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { getGiftDetail } from "@/api/bundle/api";
+import { getGiftDetail } from "@/api/my-bundle/api";
 
 export const useGiftDetailQuery = (giftId: number, bundleId: number) => {
   return useQuery({
