+++ src/queries/useMyBundlesQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { getMyBundles } from "@/api/bundle/api";
+import { getMyBundles } from "@/api/my-bundle/api";
 
 export const useMyBundlesQuery = () => {
   return useQuery({
